ref: Shorten long name files
