--- backend/app/core/settings/app_dev_settings.py
@@ -1,13 +1,0 @@
-from app.core.settings.app_settings import AppSettings
-
-
-class AppDevSettings(AppSettings):
-    """
-    A class that set up the application with devlopment environment settings.
-    """
-
-    title: str = "iWitness - Backend Development Environment Settings"
-    debug: bool = True
-
-    class Config(AppSettings.Config):
-        env_file = "env/.env.development"
