+++ backend/pyproject.toml
@@ -25,6 +25,10 @@ requests = "^2.27.1"
 asgi-lifespan = "^1.0.1"
 argon2-cffi = "^21.3.0"
 bcrypt = "^3.2.0"
+jwt = "^1.3.1"
+PyJWT = {extras = ["crypto"], version = "^2.3.0"}
+PyPika = "^0.48.9"
+python-decouple = "^3.6"
 
 [tool.poetry.dev-dependencies]
 alembic = "^1.7.7"
@@ -120,7 +124,7 @@ disable = [
   "C0103", "C0114", "C0116", "C0115", "C0330", "C0326", "C0411", "C0415",
   "E1101", "E1111", "E0213", "E0401", "E0611", "E1121",
   "R0201", "R0801", "R0903",
-  "W0106", "W0143", "W0613", "W0621", "W0622",
+  "W0106", "W0143", "W0613", "W0621", "W0622", "W1203"
 ]
 
 [tool.mypy]
